gr-limesuiteng: extract common functionality of sink/source into base block
