+++ plugins/gr-limesuiteng/python/limesuiteng/bindings/sdrdevice_sink_python.cc
@@ -14,7 +14,7 @@
 /* BINDTOOL_GEN_AUTOMATIC(0)                                                       */
 /* BINDTOOL_USE_PYGCCXML(0)                                                        */
 /* BINDTOOL_HEADER_FILE(sdrdevice_sink.h)                                        */
-/* BINDTOOL_HEADER_FILE_HASH(b2df2da9f4308e20c9cdfc4f5b5c4bbf)                     */
+/* BINDTOOL_HEADER_FILE_HASH(1a00d61ebb47b68e02d645614fe34ce2)                     */
 /***********************************************************************************/
 
 #include <pybind11/complex.h>
